+++ packages/form/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @terraprisma/form
 
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @terraprisma/general@0.0.1
+  - @terraprisma/data-display@0.0.1
+
 ## 0.0.0
 
 ### Patch Changes